requirements.txt
beautifulsoup4
deprecation
img2pdf
pikepdf
Pillow
pycnnum
requests
tqdm
